requirements_m1.txt
# works in mac m1
boto3==1.34.94
botocore==1.34.94
cffi==1.16.0
configparser==7.0.0
cryptography==42.0.5
jmespath==1.0.1
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
PyYAML==6.0.2
# instead of psycopg2
psycopg2-binary==2.9.9